- Database handling tripLimits,DailyLimits, zones and passes.

+++ shellinfo_common/src/main/java/com/shellinfo/common/data/local/db/entity/ZoneTable.kt
@@ -26,5 +26,5 @@ data class ZoneTable(
 
 
     @ColumnInfo(name = "VERSION")
-    val version: Int
+    val version: Double
 )
